+++ runtest.bat
@@ -1,0 +1,3 @@
+@echo off
+
+python -m unittest discover encryption.tests
